\name{dog}
\alias{dog}
\docType{data}
\title{Coronary sinus potassium concentration measurements for 36 dogs}
\description{
  The dogs were divided into 4 treatment groups, and the measurements for each dog
were taken every two minutes from 1 to 13 minutes after occlusion (i.e. an
artificially induced heart attack)
Taken from package assist, see there for a description.
}
\source{
Grizzle, J. E. and Allen, D. M. (1969). Analysis of growth and dose response curves, Biometrics
	25: 357-381
}
\keyword{datasets}

