\name{union}
\alias{union}
\docType{data}
\title{Union membership and hourly wages for 534 US workers}
\description{
 Union membership and hourly wages for 534 US workers.
 Taken from package SemiPar.
}
\source{
Berndt, E.R. (1991) The Practice of Econometrics. New York: Addison-Wesley.
}
\keyword{datasets}

